+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## - Unreleased
 
 - Changed handling of `NavigationService` with its rerouting, etc. events, it's moved out of `MapboxVisionAR` module, now it is in `ARContainerViewController`
-- Updated welcome message and links to Terms of Service and Privacy Policy.
+- Changed welcome message and links to Terms of Service and Privacy Policy.
 
 ## 1.0-21